flask/requirements.txt
cffi==1.15.0
click==8.0.3
colorama==0.4.4
cryptography==35.0.0
flake8==4.0.1
Flask==2.0.2
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-WTF==1.0.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
marshmallow==3.14.1
mccabe==0.6.1
mysql==0.0.3
mysql-connector-python==8.0.27
mysqlclient==2.1.0
protobuf==3.19.1
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.3.0
python-dotenv==0.19.2
six==1.16.0
Werkzeug==2.0.2
WTForms==3.0.0
